11/17

+++ src/main/java/com/lotus/booking/DTO/TranDetailResponse.java
@@ -1,11 +1,13 @@
 package com.lotus.booking.DTO;
 
+import java.math.BigDecimal;
+
 public interface TranDetailResponse {
     Long getId();
     String getName();
     Double getAmount();
     String getPayBy();
-    int getTip();
+    BigDecimal getTip();
     String getNote();
     String getDate();
     int getCount();
