+++ src/main/java/com/lotus/booking/DTO/TranResponseForDates.java
@@ -2,11 +2,13 @@ package com.lotus.booking.DTO;
 
 import com.lotus.booking.Entity.User;
 
+import java.math.BigDecimal;
+
 public interface TranResponseForDates {
     Double getAmount();
     String getDate();
     int getCount();
-    int getTip();
+    BigDecimal getTip();
 
     Long getId();
     String getFirst_Name();
