+++ src/main/java/com/lotus/booking/Entity/Transaction.java
@@ -5,6 +5,7 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
 import java.util.Date;
 @Data
 @Entity
@@ -19,7 +20,7 @@ public class Transaction {
     private String name;
     private Double amount;
     private String payBy;
-    private int tip;
+    private BigDecimal tip;
     private String note;
     private String date;
     private int count;
